Divorce CAutoRef from CAutoP [#118594399]
