+++ libgpos/include/gpos/common/CAutoRef.h
@@ -18,7 +18,7 @@
 #define GPOS_CAutoRef_H
 
 #include "gpos/base.h"
-#include "gpos/common/CAutoP.h"
+#include "gpos/common/CAutoPointerBase.h"
 #include "gpos/common/CRefCount.h"
 
 namespace gpos
@@ -32,10 +32,11 @@ namespace gpos
 	//
 	//---------------------------------------------------------------------------
 	template <class T>
-	class CAutoRef : public CAutoP<T>
+	class CAutoRef : public CAutoPointerBase<T>
 	{
 
 		private:
+			typedef CAutoPointerBase<T> _base;
 
 			// hidden copy ctor
 			CAutoRef<T>
@@ -49,14 +50,14 @@ namespace gpos
 			explicit
 			CAutoRef<T>()
 				:
-				CAutoP<T>()
+				_base()
 			{}
 
 			// ctor
 			explicit
 			CAutoRef<T>(T *pt)
 				:
-				CAutoP<T>(pt)
+				_base(pt)
 			{}
 
 			virtual ~CAutoRef();
@@ -64,7 +65,7 @@ namespace gpos
 			// simple assignment
 			CAutoRef<T> const & operator = (T* pt)
 			{
-				CAutoP<T>::m_pt = pt;
+				_base::m_pt = pt;
 				return *this;
 			}
 
@@ -81,13 +82,10 @@ namespace gpos
 	template <class T>
 	CAutoRef<T>::~CAutoRef()
 	{
-		if (NULL != CAutoP<T>::m_pt)
+		if (NULL != _base::m_pt)
 		{
-			reinterpret_cast<CRefCount*>(CAutoP<T>::m_pt)->Release();
+			reinterpret_cast<CRefCount*>(_base::m_pt)->Release();
 		}
-
-		// null out pointer before ~CAutoP() gets called
-		CAutoP<T>::m_pt = NULL;
 	}
 }
